# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 集成与管控子系统 API
  description: |
    集成与管控子系统 REST API 文档

    提供统一身份认证、安全管控、系统集成、运维监控等功能接口。

    **国产化技术栈：**
    - 数据库：达梦 DM8
    - 国密算法：SM2/SM3/SM4
    - 认证：OAuth2.0 + JWT

    **安全特性：**
    - 三权分立（管理员/审计员/操作员）
    - 数据分级管控（L1-L4）
    - 完整审计日志
  version: 1.0.0
  contact:
    name: CTG Integration Team
servers:
  - url: http://localhost:8080/api
    description: 开发环境
...
paths:
  /auth/login:
    post:
      tags: [认证管理]
      summary: 用户登录
      description: 使用用户名密码登录系统，返回JWT令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 认证失败
        '423':
          description: 账户已锁定
  /auth/refresh:
    post:
      tags: [认证管理]
      summary: 刷新令牌
      description: 使用刷新令牌获取新的访问令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/logout:
    post:
      tags: [认证管理]
      summary: 用户登出
      description: 使当前令牌失效
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 登出成功
  /auth/mfa/verify:
    post:
      tags: [认证管理]
      summary: MFA验证
      description: 验证双因子认证码
      security:
        - bearerAuth: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 验证结果
          content:
            application/json:
              schema:
                type: boolean
  /auth/password/change:
    post:
      tags: [认证管理]
      summary: 修改密码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 修改成功
  /auth/permissions:
    get:
      tags: [认证管理]
      summary: 获取权限
      description: 获取当前用户权限列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 权限列表
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /auth/validate:
    get:
      tags: [认证管理]
      summary: 验证令牌
      description: 验证JWT令牌是否有效
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 验证结果
          content:
            application/json:
              schema:
                type: boolean
...
  /audit/logs:
    get:
      tags: [审计日志]
      summary: 查询日志
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
        - name: operation
          in: query
          schema:
            type: string
        - name: startTime
          in: query
          schema:
            type: string
        - name: endTime
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 日志列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogVO'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
  /audit/statistics:
    get:
      tags: [审计日志]
      summary: 统计数据
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
        - name: startTime
          in: query
          schema:
            type: string
        - name: endTime
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 统计数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditStatistics'
  /audit/clean:
    post:
      tags: [审计日志]
      summary: 清理日志
      description: 清理过期审计日志
      security:
        - bearerAuth: []
      parameters:
        - name: retentionDays
          in: query
          schema:
            type: integer
            default: 180
      responses:
        '200':
          description: 清理数量
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /data-classification:
    post:
      tags: [数据分级]
      summary: 创建分级
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassificationRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataClassificationVO'
    get:
      tags: [数据分级]
      summary: 分级列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 分级列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataClassificationVO'
  /data-classification/{id}:
    get:
      tags: [数据分级]
      summary: 获取分级
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 分级详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataClassificationVO'
    put:
      tags: [数据分级]
      summary: 更新分级
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClassificationRequest'
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [数据分级]
      summary: 删除分级
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
  /data-classification/level/{level}:
    get:
      tags: [数据分级]
      summary: 按级别获取
      security:
        - bearerAuth: []
      parameters:
        - name: level
          in: path
          required: true
          schema:
            type: string
            enum: [L1, L2, L3, L4]
      responses:
        '200':
          description: 分级信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataClassificationVO'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        mfaCode:
          type: string
        clientIp:
          type: string
        userAgent:
          type: string
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
        username:
          type: string
        realName:
          type: string
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
        mfaRequired:
          type: boolean
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword, confirmPassword]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
          maxLength: 32
        confirmPassword:
          type: string
...
    AuditLogVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        username:
          type: string
        realName:
          type: string
        clientIp:
          type: string
        operation:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
        resourceName:
          type: string
        description:
          type: string
        requestMethod:
          type: string
        requestUrl:
          type: string
        responseStatus:
          type: integer
        result:
          type: string
          enum: [SUCCESS, FAILURE]
        duration:
          type: integer
          format: int64
        dataLevel:
          type: string
        createdAt:
          type: string
          format: date-time
    AuditStatistics:
      type: object
      properties:
        totalOperations:
          type: integer
          format: int64
        successCount:
          type: integer
          format: int64
        failureCount:
          type: integer
          format: int64
        operationCounts:
          type: array
          items:
            type: object
            properties:
              operation:
                type: string
              count:
                type: integer
                format: int64
    CreateClassificationRequest:
      type: object
      required: [classificationName, securityLevel]
      properties:
        classificationName:
          type: string
        securityLevel:
          type: string
          enum: [L1, L2, L3, L4]
        requireEncryption:
          type: boolean
        requireMasking:
          type: boolean
        maskingRule:
          type: string
        accessPolicy:
          type: string
        description:
          type: string
    UpdateClassificationRequest:
      type: object
      properties:
        classificationName:
          type: string
        requireEncryption:
          type: boolean
        requireMasking:
          type: boolean
        maskingRule:
          type: string
        accessPolicy:
          type: string
        description:
          type: string
    DataClassificationVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        classificationName:
          type: string
        securityLevel:
          type: string
        requireEncryption:
          type: boolean
        requireMasking:
          type: boolean
        maskingRule:
          type: string
        accessPolicy:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...